requirements.txt
python-binance
TA-Lib
numpy
pandas
websocket_client
flask
flask_cors
backtrader
matplotlib==3.2.2